+++ src/day8/re_frame_10x/fx/local_storage.cljs
@@ -4,8 +4,7 @@
     [goog.storage.mechanism.HTML5LocalStorage]
     [cljs.reader                                                  :as reader]
     [clojure.string                                               :as string]
-    [day8.re-frame-10x.inlined-deps.re-frame.v1v1v2.re-frame.core :as rf])
-  (:refer-clojure :exclude [get]))
+    [day8.re-frame-10x.inlined-deps.re-frame.v1v1v2.re-frame.core :as rf]))
 
 (def storage (goog.storage.Storage. (goog.storage.mechanism.HTML5LocalStorage.)))
 
@@ -15,20 +14,15 @@
   "Adds a unique prefix to local storage keys to ensure they don't collide with the host application"
   (str safe-prefix key))
 
-(defn get
-  "Gets a re-frame-10x value from local storage."
+(defn load
+  "Loads a re-frame-10x value from local storage."
   ([key]
-   (get key nil))
+   (load key nil))
   ([key not-found]
    (let [value (.get storage (safe-key key))]
      (if (undefined? value)
        not-found
        (reader/read-string value)))))
-
-(defn save!
-  "Saves a re-frame-10x value to local storage."
-  [key value]
-  (.set storage (safe-key key) (pr-str value)))
 
 (defn delete-all-keys!
   "Deletes all re-frame-10x config keys"
@@ -37,23 +31,23 @@
     (when (string/starts-with? k safe-prefix)
       (.remove storage k))))
 
-(defn after
+(defn save
   ([key]
    (rf/after
-     (fn [x]
-       (save! key x))))
+     (fn [db]
+       (.set storage (safe-key key) (pr-str db)))))
   ([key & ks]
    (rf/after
-     (fn [x]
+     (fn [db]
        (run!
          (fn [k]
-           (let [v (if (vector? k) (get-in x k) (get x k))]
-             (save! key v)))
+           (let [v (if (vector? k) (get-in db k) (get db k))]
+             (.set storage (safe-key key) (pr-str v))))
          ks)))))
 
 (rf/reg-cofx
-  ::get
+  ::load
   (fn [coeffects {:keys [key or]}]
     (assoc coeffects
       (keyword key)
-      (get key or))))
+      (load key or))))